ABP (int) and IQ(short) now carried through from DMA to RR314. Need to add logic which keeps track of the expected number of samples, and fills the RR314 buffers appropriately.
git-svn-id: http://svn.eol.ucar.edu/svn/eldora/trunk@184 15f24ed5-0439-0410-88a4-bd1d0fc06cdd

+++ redrapids/RR314sim.h
@@ -49,9 +49,9 @@ namespace RedRapids {
     /// The channel numbers which receive ABP data.
     std::vector<int> _abpChans;
     /// A buffer used as the source of ABP data.
-    std::vector<unsigned int> _abp;
+    std::vector<int> _abp;
     /// A buffer used as the source of IQ data.
-    std::vector<unsigned int> _iq;
+    std::vector<short> _iq;
     /// The number of gates
     int _gates;
     /// The starting iq gate
